Implement a binary tree algorithm maze generator
This is a simple and quick algorithm I got from the book "Mazes for
Programmers" by Jamis Buck. I do not use his Grid and Cell classes
because I needed to quickly integrate it with the existing 2d vector.

The 2D vector contains both the cells and walls.

Because the generator steps by 2 the length of the map sides need to be
odd. This should be fixed in the future.